Hopefully Final Set of Tweaks and Fixes
Increased the maximum Evasion that a ship can have to 95% from 90%.
Fixed the occasional buggy behavior of Picket Combat Computers.
Tweaked various Faction demands and attractions to make more sense and be less annoying.
Large batch of ship designer fixes to get the ship designer to not do weird stuff.